+++ qadence/analog/__init__.py
@@ -1,5 +1,5 @@
 from __future__ import annotations
 
-from .add_interaction import add_interaction
+from .interaction import add_interaction
 
 __all__ = ["add_interaction"]
